"Send Crash Report User Confirmation" Dialog after restart (iOS)
SetUserConfirmationHander method added to Crashes SDK.
Also Puppet shows confirmation dialog after crash and restart using this method.
Say Crashes.SetUserConfirmationHandler(UserConfirmationHandler) in Awake() method to add handler.
Crashes service starts not in AppCenterStarter.original, but later, in AppCenterBehaviour->StartAppCenter(), to give app an opportunity to set handler in Awake()
For now array of error reports is not passed to user confirmation handler. It can be implemented later if needed.